requirements.txt
flask_uploads
alembic
flask_restful
passlib
Babel
blinker
certifi
chardet
click
dominate
elasticsearch
Flask
Flask-Babel
Flask-Bootstrap
Flask-HTTPAuth
Flask-Login
Flask-Mail
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess_language-spirit==0.5.3
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
PyJWT==1.5.3
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2017.2
redis==2.10.6
requests==2.18.4
rq==0.9.2
six==1.11.0
SQLAlchemy
urllib3
visitor
Werkzeug
WTForms

# requirements for Heroku
psycopg2-binary
gunicorn
